+++ scripts/lib/namedtuple.py
@@ -1,0 +1,4 @@
+from collections import namedtuple
+
+def dict_to_namedtuple(name: str, dic: dict):
+  return namedtuple(name, dic.keys())(*dic.values())
